[Feature] Update packets to 2022 (#112)
* Force sync load on loading screens

* Change the way escape window is built

* Create packet log window

* Create option to connect to localhost directly

* ACCEPT_ENTER

* ACCEPT_MAKECHAR

* CZ.ENTER2

* ZC_FRIENDS_LIST

* Inventory packets

* ZC_REPUTE_INFO

* attempt to fix scroll view

* ZC_PURCHASE_ITEMLIST

* ITEM_PICKUP_ACK

* dirty fix on inventory

* stop awaiting stuff at game manager start

* ZC_ATTACK_FAILURE_FOR_DISTANCE

* ZC_NOTIFY_EFFECT & ZC_USE_SKILL

* Fix packet name and PACKET_ZC_FRIENDS_LIST_sub block size

* change packet log window

+++ UnityClient/Assets/Scripts/Renderer/Entities/EntityCanvas.cs
@@ -39,13 +39,13 @@ public class EntityCanvas : MonoBehaviour {
         (EntityName.gameObject.transform as RectTransform).sizeDelta = textSize;
     }
 
-    public void SetEntityHP(int value, int maxValue) {
+    public void SetEntityHP(long value, long maxValue) {
         HPBar.minValue = 0f;
         HPBar.maxValue = maxValue;
         HPBar.value = value;
     }
 
-    public void SetEntitySP(int value, int maxValue) {
+    public void SetEntitySP(long value, long maxValue) {
         SPBar.minValue = 0f;
         SPBar.maxValue = maxValue;
         SPBar.value = value;
